split the code into modules, task 4
