+++ index.html
@@ -228,5 +228,7 @@
     </div>
   </template>
 
+  <!-- Подключение скрипта -->
+  <script async src="./js/main.js">alert(123)</script>
 </body>
 </html>
