+++ js/data.js
@@ -1,0 +1,43 @@
+const CHEKIN = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const CHEKOUT = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+
+const DESCRIPTIONS = [
+  'просторная светлое помещение',
+  'не очень просторное и темное помещение',
+  'компактно и светло',
+  'большая площадь территории',
+  'просторный двор',
+];
+
+export {CHEKIN, CHEKOUT, FEATURES, PHOTOS, TYPES, DESCRIPTIONS};
